+++ README.md
@@ -158,3 +158,12 @@ odoo pgactivity
 |NAMED_ODOO_POSTGRES_VOLUME| Use a specific external volume; not dropped with down -v command|
 |CRONJOB_DADDY_CLEANUP=0 */1 * * * ${JOB_DADDY_CLEANUP}|Turn on grandfather-principle based backup|
 |RESTART_CONTAINERS=1|Sets "restart unless-stopped" policy|
+
+
+# Pytests
+
+Best executed with:
+
+```bash
+time sudo -E pytest
+```
