+++ wodoo/exceptions.py
@@ -1,0 +1,3 @@
+
+class NoProjectNameException(Exception):
+    pass
