+++ wodoo/lib_control.py
@@ -1,6 +1,6 @@
 import click
 import os
-from . import cli, pass_config, Commands
+from .cli import cli, pass_config, Commands
 from .lib_clickhelpers import AliasedGroup
 from .tools import execute_script
 import subprocess
