+++ wodoo/lib_docker_registry.py
@@ -10,7 +10,7 @@ from datetime import datetime
 import os
 import click
 from .tools import __dc
-from . import cli, pass_config
+from .cli import cli, pass_config
 from .lib_clickhelpers import AliasedGroup
 from .tools import split_hub_url, abort
