+++ wodoo/tests/module_respartner_dummyfield2/partnerview.xml
@@ -5,7 +5,7 @@
 			<field name="inherit_id" ref="base.view_partner_form"/>
 			<field name="arch" type="xml">
 				<field name="name" position="after">
-					<field name="dummy1"/>
+					<field name="dummy2"/>
 				</field>
 			</field>
 		</record>
